Remove unnecessary type casting in channel decode call

+++ rust/oxide-decoders/src/decoders/acars.rs
@@ -425,7 +425,7 @@ pub struct ACARSDecoder {
 }
 
 impl Decoder for ACARSDecoder {
-    fn decode(&mut self, length: u32) {
+    fn decode(&mut self, length: usize) {
         self.demod_msk(length);
     }
 
@@ -477,10 +477,10 @@ impl ACARSDecoder {
         }
     }
 
-    pub fn demod_msk(&mut self, len: u32) {
+    pub fn demod_msk(&mut self, len: usize) {
         /* MSK demod */
 
-        for in_ in &mut self.dm_buffer.into_iter().take(len as usize) {
+        for in_ in &mut self.dm_buffer.into_iter().take(len) {
             let s: f32 = 1800.0 / INTRATE as f32 * 2.0 * std::f32::consts::PI + self.msk_df;
             let mut v: Complex<f32> = num::Complex::new(0.0, 0.0);
             let mut o: f32;
